requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
anyio==4.9.0
argcomplete==3.6.2
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.38.43
botocore==1.38.43
Brotli==1.1.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
choreographer==1.0.9
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dacite==1.7.0
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
executing==2.2.0
fiftyone==1.6.0
fiftyone-brain==0.21.2
fiftyone_db==1.1.7
flatbuffers==25.2.10
fonttools==4.58.4
ftfy==6.3.1
future==1.0.0
gast==0.6.0
glob2==0.7
google-pasta==0.2.0
graphql-core==3.2.6
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
inflate64==1.0.3
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kaleido==1.0.0
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
logistro==1.1.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mongoengine==0.29.1
motor==3.6.1
multivolumefile==0.2.3
namex==0.1.0
narwhals==1.44.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.1.3
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pprintpp==0.4.0
priority==2.0.0
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
py7zr==1.0.0
pybcj==1.0.6
pycryptodomex==3.23.0
pydash==8.0.5
Pygments==2.19.2
pymongo==4.9.2
pyparsing==3.2.3
pyppmd==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
pyzstd==0.17.0
rarfile==4.2
regex==2024.11.6
requests==2.32.4
retrying==1.4.0
rich==14.0.0
rtree==1.4.0
s3transfer==0.13.0
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spyder-kernels==3.0.5
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.47.1
strawberry-graphql==0.275.2
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tkinterdnd2==0.4.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
universal-analytics-python3==1.1.1
urllib3==2.5.0
voxel51-eta==0.14.2
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2